refactor: merge config loader into a single config factory
