--- src/config/jwt.ts
@@ -1,21 +1,0 @@
-import { registerAs } from '@nestjs/config';
-
-import { validateEnv } from '@/utils/helpers/envs';
-
-export interface IJwtConfig {
-  secret: string;
-  expiresIn: string;
-  refreshSecret: string;
-  refreshExpiresIn: string;
-}
-
-export const jwtConfig = registerAs('jwt', (): IJwtConfig => {
-  const env = validateEnv();
-
-  return {
-    secret: env.JWT_SECRET,
-    expiresIn: env.JWT_EXPIRES_IN,
-    refreshSecret: env.JWT_REFRESH_SECRET,
-    refreshExpiresIn: env.JWT_REFRESH_EXPIRES_IN,
-  };
-});
