+++ src/shared/utils/hash.spec.ts
