+++ src/shared/utils/hash.ts
